+++ lib/zaikio/directory/revoked_access_token.rb
@@ -1,7 +1,7 @@
 module Zaikio
   module Directory
     class RevokedAccessToken < Base
-      uri "blacklisted_access_tokens"
+      uri "revoked_access_tokens"
     end
   end
 end